improve up.sh

+++ up.sh
@@ -1,10 +1,14 @@
 #!/usr/bin/env bash
 
+if [[ `git status --porcelain --untracked-files=no` ]]; then
+	echo untracked files, aborting...
+	exit 1
+fi
+
 OLDVERSION=$(jq -r '.version' ./package.json)
 read -p "New version (old was $OLDVERSION): " NEWVERSION
 
 rm -rf lib/
-git stash save 'up.sh auto-stash'
 sed -i "s/$OLDVERSION/$NEWVERSION/g" package.json
 git commit -S -a -m "up to $NEWVERSION" && \
 	git tag -s "$NEWVERSION" -m "$NEWVERSION" && \
@@ -12,7 +16,6 @@ git commit -S -a -m "up to $NEWVERSION" && \
 	git push --tags && \
 	npm pub
 rm -rf lib/
-git stash pop
 git reset .
 
 # update docs
